% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBam.R
\name{bamByFeatures}
\alias{bamByFeatures}
\title{filtering a bam file by feature information}
\usage{
bamByFeatures(bam, readLen = 51, gtf, chr = "chr6", feature = "gene")
}
\arguments{
\item{bam}{a bam file}

\item{readLen}{a size of a single read}

\item{gtf}{gtf file}

\item{chr}{chromosome}

\item{feature}{feature type, currently filtering with either 'gene' or 'transcript' feature is available.}
}
\description{
It deals with only singe chromosome so a bam file should be filtered by one chromosome (with a certain location). 
Using a filtered gtf with single chromosome or a certain region of the chromosome will make it run faster.
}
