% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDfromBams.R
\name{EDfromBams}
\alias{EDfromBams}
\title{To get editing distance from multiple bam files}
\usage{
EDfromBams(bamFiles, gtf, output_name = "ed_table.txt", virtualenv = 0,
  r = "VARSL-VARS2,C6orf205-MUC21")
}
\arguments{
\item{bamFiles}{bam files <path/to/BAM_file_1> <path/to/BAM_file_2>}

\item{gtf}{gtf file <path/to/GTF_file>}

\item{output_name}{output name}

\item{virtualenv}{path of virtualenv, optional}

\item{r}{combine redundnat gene names, example, VARSL-VARS2,C6orf205-MUC21, optional}
}
\description{
To get editing distance from multiple bam files
}
\examples{
bamFiles=system.file("extdata", "mapping2*bam", package = "AltHapAlignR")
gtf=system.file("extdata", "gencode.v21.chr_patch_hapl_HLA.annotation.gtf", 
                 package = "AltHapAlignR")
EDfromBams(bamFiles, gtf, output_name="output.txt", virtualenv=0, r="VARSL-VARS2,C6orf205-MUC21")
}
