% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{fasta2DNAStringSet}
\alias{fasta2DNAStringSet}
\title{fasta2DNAStringSet}
\usage{
fasta2DNAStringSet(fasta)
}
\arguments{
\item{fasta}{fastq format data}
}
\description{
converting fasta format to DNAStringSet
}
