% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{filterAndMergeExons}
\alias{filterAndMergeExons}
\title{Filter and merge a exon table}
\usage{
filterAndMergeExons(exon_table, seqname = "chr6", levelBelow = 3,
  transcript_type = "protein_coding")
}
\arguments{
\item{exon_table}{output from 'gtf2ExonTable'}

\item{seqname}{chromosome name, it uses only single chromosome}

\item{levelBelow, }{evidence level of annotation. If levelBelow=3, it uses only level 1 and 2 data. levels: 1(verified loci), 2(manually annotated loci), 3(automatically annotated loci)}

\item{transcript_type}{filter exons by specific biotypes. transcript_type="all" uses all data. default: 'protein_coding', Further details about transcript biotypes are here. http://www.gencodegenes.org/gencode_biotypes.html}
}
\description{
This filers exon data by annotation level and merges exons across multiple isoforms.
}
\examples{
# to use your own reference genome
gtf=system.file("extdata", "gencode.v21.hla.annotation.gtf", package = "HapTag")
exTab <- gtf2ExonTable(gtf, refver="hg38", grange=c("chr6", "28528216", "33389373"))
filtered <- filterAndMergeExons(exTab)

# to use HapTagAnno containing human reference genomes and parsed data.
library(HapTagAnno.hg38)
gtf <- system.file("extdata/gencode.v21.hla.annotation.gtf", package="HapTag")
exTab <- gtf2ExonTable(gtf, refver="hg38", grange=c("chr6", "28528216", "33389373"))
filtered <- filterAndMergeExons(exTab)

}
