% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{filterExons}
\alias{filterExons}
\title{To filter exon feature from a GTF file}
\usage{
filterExons(gtf, type = "protein_coding", max_level = 2, status = "KNOWN")
}
\arguments{
\item{gtf}{GTF file}

\item{type}{selecting exon feature}

\item{max_level}{annotation level}

\item{status}{exon status}
}
\description{
To filter exon feature from a GTF file
}
\examples{
output <- filterExons(gtf, type="protein_coding", max_level=2, status="KNOWN")

# Output data
# summary of exon information, format: data.frame
}
