% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{filterGenes}
\alias{filterGenes}
\title{To filter gene feature from a GTF file}
\usage{
filterGenes(gtf, type = "protein_coding", max_level = 2, status = "KNOWN")
}
\arguments{
\item{gtf}{GTF file}

\item{type}{selecting gene feature}

\item{max_level}{annotation level}

\item{status}{gene status}
}
\description{
To filter gene feature from a GTF file
}
\examples{
output <- filterGenes(gtf, type="protein_coding", max_level=2, status="KNOWN")

# Output data
# summary of gene information, format: data.frame
}
