% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{getExons}
\alias{getExons}
\title{get exon data}
\usage{
getExons(gtf_granges)
}
\arguments{
\item{gtf_granges}{GRanges object}
}
\description{
It produces min.info <- c("gene_id", "transcript_id", "gene_type", "gene_status", "gene_name", "transcript_type", "transcript_status",  "transcript_name", "exon_number", "exon_id", "level")
}
