% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDfromBams.R
\name{getGeneList}
\alias{getGeneList}
\title{To get gene feature from a GTF file}
\usage{
getGeneList(gtf, type = "protein_coding")
}
\arguments{
\item{gtf}{GTF file}

\item{type}{selecting gene feature}
}
\description{
To get gene feature from a GTF file
}
\examples{
gtf=system.file("extdata", "gencode.v21.chr_patch_hapl_HLA.annotation.gtf", 
                 package = "AltHapAlignR")
output <- getGeneList(gtf, type="protein_coding")

# Output data
# summary of gene information, format: data.frame
}
