% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{getHapsByShortestPaths}
\alias{getHapsByShortestPaths}
\title{To predict shortest path}
\usage{
getHapsByShortestPaths(paired_mapping_rates, anno = "hg38", penalty = 1,
  sample_name)
}
\arguments{
\item{paired_mapping_rates}{output 'predictRecombHapByGenes'  (*max_comb.txt)}

\item{anno}{gtf version (ex, "hg38")}

\item{penalty}{1: penalty applied, 0: penalty not applied}

\item{sample_name}{output file name}
}
\description{
To predict shortest path
}
\examples{
output <- getHapsByShortestPaths(paired_mapping_rates, anno="hg38", penalty=1, sample_name)

# Output data
# *.shortestPath_heatmap.pdf : heatmap of selected haplotypes
# gene and haplotype to draw heatmap
heatmap_genes <- output$heatmap_genes
# predicted best pairs of haploytpes
summary_of_best_pairs <- output$summary_of_best_pairs
}
