% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{getHapsByShortestPaths_withPermutation_test}
\alias{getHapsByShortestPaths_withPermutation_test}
\title{To predict shortest path using results by permutation test}
\usage{
getHapsByShortestPaths_withPermutation_test(paired_mapping_rates,
  permutation_result, anno = "hg38", penalty = 1, sample_name)
}
\arguments{
\item{paired_mapping_rates}{output of 'predictRecombHapByGenes'  (*max_comb.txt)}

\item{permutation_result}{output of 'permutation_test'  (*permutation_test.txt)}

\item{anno}{gtf version (ex, "hg38")}

\item{penalty}{1: penalty applied, 0: penalty not applied}

\item{sample_name}{output file name}
}
\description{
To predict shortest path using results by permutation test
}
\examples{
output <- getHapsByShortestPaths_withPermutation_test(paired_mapping_rates, anno="hg38", penalty=1, sample_name)

# Output data
# *.shortestPath_heatmap.pdf : heatmap of selected haplotypes
# predicted best pairs of haploytpes
summary_of_best_pairs <- output$summary_of_best_pairs
}
