% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{getMappingRatesFromPairs}
\alias{getMappingRatesFromPairs}
\title{To get mapping rates in all pairs of haplotypes and produce summerized figure}
\usage{
getMappingRatesFromPairs(ed, editing_distance = 10, minDP_per_gene = 30,
  anno = "hg38", sample_name, read_length = 50)
}
\arguments{
\item{ed}{output of 'predictRecombHapByGenes'}

\item{editing_distance}{number of mismatch}

\item{anno}{gtf version (ex, hg38)}

\item{sample_name}{output file name}

\item{read_length}{read length in bam files}

\item{minDP_per_minimum}{read depth in genes}
}
\description{
To get mapping rates in all pairs of haplotypes and produce summerized figure
}
\examples{
output <- getMappingRatesFromPairs(ed, editing_distance=10, minDP_per_gene=30, anno="hg38", sample_name, read_length=50)

# Output data
# *mhc_heatmap.pdf : heatmap of selected haplotypes, a summerized figure of haplotype-specific mapping
# mapping rates and gene counts in all pairs of haplotypes
max_comb <- output$max_comb
# Mapping ranks in all pairs of haplotypes
mapping_ranks <- output$selected_genes
}
