% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{getOverlappingExons}
\alias{getOverlappingExons}
\title{To get overlapping exons in different genes}
\usage{
getOverlappingExons(overlappingTrx, anno_exon)
}
\arguments{
\item{overlappingTrx}{output from getOverlappingTrx}

\item{anno_exon}{output from filterExons}
}
\description{
To get overlapping exons in different genes
}
\examples{
output <- getOverlappingExons(overlappingTrx, anno_exon)

# Output data
# overlapping exons in different genes
}
