% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{getOverlappingTrx}
\alias{getOverlappingTrx}
\title{To get overlapping transcripts in genes}
\usage{
getOverlappingTrx(gene_cordi)
}
\arguments{
\item{gene_cordi}{data.table format (chr, gene_name, start, end)}
}
\description{
To get overlapping transcripts in genes
}
\examples{
output <- getOverlappingTrx(gene_cordi)

# Output data
# overlapping transcripts in genes
}
