% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHapExons.R
\name{getVariantsByAlignment}
\alias{getVariantsByAlignment}
\title{get variants between two sequences after pairwiseAlignment}
\usage{
getVariantsByAlignment(refver = "hg38", refseq, hap1_dbSNP, hap2_dbSNP, seq1,
  seq2, hap_names = c("hap1", "hap2"))
}
\arguments{
\item{refver}{a version of human genome reference (it may not need here...)}

\item{refseq}{reference sequences. If no reference sequence is supplied the \code{ref_MHC} object
from the \code{HapTagAnno} package that was last loaded will be used.}

\item{hap1_dbSNP}{dbSNP data of haplotype 1, should be loaded before running the function}

\item{hap2_dbSNP}{dbSNP data of haplotype 2, should be loaded before running the function}

\item{seq1}{position information of first sequence, c(chromsome name, start position, end position, strand)}

\item{seq2}{position information of second sequence, c(chromsome name, start position, end position, strand)}

\item{hap_names}{names of two haplotypes}
}
\description{
get variants between two sequences after pairwiseAlignment
}
\examples{
# load reference sequences and  
library(HapTagAnno.hg38)
# load snp data
hap1_dbSNP <- subset(dbSNP_MHC, dbSNP_MHC$chr=="chr6")
hap2_dbSNP <- subset(dbSNP_MHC, dbSNP_MHC$chr=="chr6_GL000251v2_alt")
# load reference sequences and dbSNP by a function 'loadDataForVariantsByAlignment'
loadDataForVariantsByAlignment(refver="hg38", seq1="chr6", seq2="GL000251.2")

}
