% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHapExons.R
\name{getVariantsByAlignment__}
\alias{getVariantsByAlignment__}
\title{get variants between two aligned sequences}
\usage{
getVariantsByAlignment__(alignment, snp)
}
\arguments{
\item{alignment}{alignment}

\item{snp}{snp}
}
\description{
get variants between two aligned sequences
}
