% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{getWeightReads}
\alias{getWeightReads}
\title{Weighting reads by editing distance to reference sequences}
\usage{
getWeightReads(read_to_gene, minDP_per_gene = 30)
}
\arguments{
\item{read_to_gene}{output from 'predictRecombHapByGenes' (*ed_all.txt)}

\item{minDP_per_gene}{minimum read depth required in gene}

\item{bamFiles}{bam files}
}
\description{
Weighting reads by editing distance to reference sequences
}
\examples{
output <- getWeightReads(read_to_gene, minDP_per_gene=30)

# Output data
# mapping rates in each haploypte
mapping_stat <- output$ed_summary
# gene counts in haploytpes
uniqueFragmentID <- output$mapping_stats_to_genes
# unique mapping rates in all pairs of haploytpes
list_pairedReads <- output$rates_of_weighed_reads_by_genes
}
