% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{gtf2ExonTable}
\alias{gtf2ExonTable}
\title{Convert a GTF file to an exon table}
\usage{
gtf2ExonTable(gtf, refver = "hg38", grange = c("chr6", "28524925",
  "33389373"))
}
\arguments{
\item{gtf}{Transcript annotations. Can either be provided as a GRanges object or as the name of
a GTF file containing the annotations.}

\item{refver}{human genome reference version}

\item{grange}{Character vector of the form \code{c(chromosome, start, end)}, 
indicating the genomic region from which exons should be considered. Set \code{start}
or \code{end} to 0 to indicate that the region should extend to the respective end 
of the chromosome. Setting this to \code{NULL} will disable subsetting altogether.}
}
\description{
This will produce a table of parsed exon data from gtf file. It uses a function 'getExons'.
}
\examples{
gtf=system.file("extdata", "gencode.v21.hla.annotation.gtf", package = "HapTag")
gtf2ExonTable(gtf, refver="hg38", grange=c("chr6", "28528216", "33389373") )

# to use HapTagAnno containing human reference genomes and parsed data.
library(HapTagAnno.hg38)
gtf2ExonTable(gtf, refver="hg38", grange=c("chr6", "28528216", "33389373") )
}
