% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{heatmapByShortestPaths}
\alias{heatmapByShortestPaths}
\title{To predict shortest path}
\usage{
heatmapByShortestPaths(paired_mapping_rates, gtf, hap_names = c("apd", "cox",
  "dbb", "mann", "mcf", "pgf", "qbl", "ssto"), penalty = 1,
  sample_name = "sample_name")
}
\arguments{
\item{paired_mapping_rates}{output from 'predictRecombHapByGenes'}

\item{gtf}{a gtf file name containing gene feature of all haplotypes}

\item{hap_names}{haplotype names matching to column names (eg. c("apd", "cox", "dbb", "mann", "mcf", "pgf", "qbl", "ssto") )}

\item{penalty}{1: penalty applied, 0: penalty not applied}

\item{sample_name}{output file name}
}
\description{
To predict shortest path
}
