% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHapExons.R
\name{loadDataForVariantsByAlignment}
\alias{loadDataForVariantsByAlignment}
\title{load snp data to find variants by aligned sequences (need to change to load dbSNP for only two haplotypes, not entire dbSNP)}
\usage{
loadDataForVariantsByAlignment(refver = "hg38", seq1 = "chr6",
  seq2 = "GL000251.2")
}
\arguments{
\item{refver}{a version of reference sequence, hg38 and hg19 versions are available from HapTagAnno.hg38 and HapTagAnno.hg19, respectively.}

\item{seq1}{a chromosome name of sequence 1}

\item{seq2}{a chromosome name of sequence 2}
}
\description{
load snp data to find variants by aligned sequences (need to change to load dbSNP for only two haplotypes, not entire dbSNP)
}
\examples{
# for reference hg38
require(HapTagAnno.hg38)
# for PGF and COX haplotypes
loadDataForVariantsByAlignment(refver="hg38", seq1="chr6", seq2="GL000251.2")

# for reference hg19
\dontrun{
require(HapTagAnno.hg19)
# for PGF and COX haplotypes
loadDataForVariantsByAlignment(refver="hg19", seq1="chr6", seq2="GL000251.1")
}
}
