% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTags.R
\name{mgsub}
\alias{mgsub}
\title{Replacing multiple patterns}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{Character vector containing regular expressions.}

\item{replacement}{Character vector containing one replacement string for each 
regular expression in \code{pattern}.}

\item{x}{Character vector where matches are sought.}

\item{...}{Further arguments passed to gsub.}
}
\description{
Allow the use of multiple pattern/replacement pairs in a single function call.
}
