% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{permutation_test}
\alias{permutation_test}
\title{To read multiple bam files}
\usage{
permutation_test(hapNames, weighted_gene, p_number = 1e+05)
}
\arguments{
\item{hapNames}{haploytpe names}

\item{weighted_gene}{output of 'predictRecombHapByGenes' (*rates_of_weighed_reads_by_genes.txt)}

\item{p_number}{permutation number}
}
\description{
To read multiple bam files
}
\examples{
output <- permutation_test(hapNames, weighted_gene, p_number=100000)

# Output data
# P-values of mapping results in all pairs of haplotypes by permutaion
}
