% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{predictRecombHapByGenes}
\alias{predictRecombHapByGenes}
\title{predictRecombHapByGenes: To predict if a sample is  heterogeneous or homogeneous AND to predic haplotypes if the sample is heterogeneous}
\usage{
predictRecombHapByGenes(bamFiles, bamChrs, bamNames,
  sample_name = "sample_name", gtf_MHC, anno_type = "protein_coding",
  anno_level = 2, anno_status = "KNOWN", minDP_per_gene = 30,
  numWorkers = 2, uniq_mapping = 1, editing_distance = 0,
  read_length = 50, save_mapping_info = 1, mapper = "tophat")
}
\arguments{
\item{bamFiles}{bam files}

\item{bamChrs}{haplotype names}

\item{bamNames}{bam file names}

\item{sample_name}{output file name}

\item{gtf_MHC}{gtf file}

\item{anno_type}{feature of gene (ex, "All", protein_coding", "processed_transcript", etc.)}

\item{anno_level}{feature level (ex, 1, 2, or 3)}

\item{anno_status}{annotation status of gene feature (ex, "KNOWN", "NOVEL", or "PUTATIVE")}

\item{minDP_per_gene}{minimum read depth in genes}

\item{numWorkers}{the number of nodes}

\item{uniq_mapping}{to get only uniquely mapped read}

\item{editing_distance}{number of mismatch}

\item{read_length}{read length in bam files}

\item{save_mapping_info}{1 for saving all mapping data, it will produce big files.}

\item{mapper}{output from only tophat availble}
}
\description{
predictRecombHapByGenes: To predict if a sample is  heterogeneous or homogeneous AND to predic haplotypes if the sample is heterogeneous
}
\examples{
# Input data: bam files from mapping to two haplotype references  
output <- predictRecombHapByGenes(bamFiles, bamChrs, bamNames, sample_name="sample_name", gtf_MHC, anno_type="protein_coding", anno_level=2, anno_status="KNOWN", minDP_per_gene=30, numWorkers=2, uniq_mapping=1, editing_distance=0, read_length=50, save_mapping_info=0, mapper="tophat")

# Output data
# *mhc_heatmap.pdf : heatmap of selected haplotypes
# *max_comb.txt : mapping rates and gene counts in all pairs of haplotypes
# *selected_genes.txt : Mapping ranks in all pairs of haplotypes
# *mapping_stats_to_genes.txt : gene counts in haployptes
# *rates_of_weighed_reads_by_genes.txt :unique  mapping rates in all pairs of haplotypes
}
