% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{readBamFiles}
\alias{readBamFiles}
\title{To read multiple bam files}
\usage{
readBamFiles(bamFiles, bamNames, mapper = "tophat")
}
\arguments{
\item{bamFiles}{bam files}

\item{bamNames}{sample names of each bam file}

\item{mapper}{mapping program used, only tophat format available}
}
\description{
To read multiple bam files
}
\examples{
output <- readBamFiles(bamfiles, name_of_bamfiles, mapper="tophat")

# Output data
# mapping statistics
mapping_stat <- output$mapping_stat
# read ids
uniqueFragmentID <- output$uniqueFragmentID
# paired reads
list_pairedReads <- output$list_pairedReads
}
