% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHapExons.R
\name{seqCompare}
\alias{seqCompare}
\title{pairwise sequence alignment}
\usage{
seqCompare(hap1_exon, hap2_exon, chrs = c("chr6", "GL000251.2"), refseq)
}
\arguments{
\item{hap1_exon}{an exon data.frame, an output from a function 'filterAndMergeExons', which column names are c("exn", "start", "end", "width", "gene_strand", "gene_name", "n_of_merged_exons", "merged_exon_ids")}

\item{hap2_exon}{another exon data.frame to compaire with hap1_exon table}

\item{chrs}{chromosome names of two haplotypes to compare}

\item{refseq}{'DNAStringSet' of reference sequences,}
}
\description{
It produce best hits of exon sequences between two haplotypes by pairwiseAlignment
}
\examples{
# to use fasta file from 'HapTagAnno.hg38'
library(HapTagAnno.hg38)
fasta=ref_MHC   # human reference version hg38

# to use your own reference genome
fasta="/path/to/custom/reference/genome.fa"

gtf <- system.file("extdata", "gencode.v21.hla.annotation.gtf", package = "HapTag")
gtf2ExonTable(gtf, refver="hg38", grange=c("chr6", "28528216", "33389373"))
}
