% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTags.R
\name{tagReads}
\alias{tagReads}
\title{To assign haplotype-specific tags}
\usage{
tagReads(hap1, hap2, mapper = "tophat")
}
\arguments{
\item{hap1}{a bam file of haplotype 1}

\item{hap2}{a bam file of haplotype 2}

\item{mapper}{mapping program used, only tophat or bwa format available}
}
\description{
To assign haplotype-specific tags
}
\examples{
# Input data: bam files from mapping to two haplotype references  
hap1 <- system.file("extdata", "MHC_PGF.bam", package = "HapTag")
hap2 <- system.file("extdata", "MHC_COX.bam", package = "HapTag")
output <- tagReads(hap1, hap2, mapper="tophat")

# Output data
# unique number of all reads and paired-end reads
stat_uniqNumber <- output$stats$uniqueNumber
# Frequency of reads for each group 
stat_groups <- output$stats$groups
# Frequency of reads for each tag
stat_tagXG <- output$stats$tagXG
# Fragment ids with mapping tags and assigned tags.
fragmentIDs <- output$tagXG
}
