% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHaps.R
\name{getMappingRatesFromPairs}
\alias{getMappingRatesFromPairs}
\title{To get mapping rates in all pairs of haplotypes and produce summerized figure}
\usage{
getMappingRatesFromPairs(ed_table, hap_names = c("apd", "cox", "dbb", "mann",
  "mcf", "pgf", "qbl", "ssto"), read_length = 50)
}
\arguments{
\item{ed_table}{a file with a table of editing distances generated from a function 'EDframBams'}

\item{hap_names}{haplotype names matching to column names of the input file 'ed_table' (eg. c("apd", "cox", "dbb", "mann", "mcf", "pgf", "qbl", "ssto") )}

\item{read_length}{read length in bam files}
}
\description{
To get mapping rates in all pairs of haplotypes and produce summerized figure
}
\examples{
ed_table=system.file("extdata", "example_ed_table.txt", package = "AltHapAlignR")
ed_table <- read.table(ed_table, sep="\\t", header=TRUE)
output <- getMappingRatesFromPairs(ed_table, 
               hap_names=c("apd", "cox", "dbb", "mann", "mcf", "pgf", "qbl", "ssto"), 
               read_length=50)
}
